--- src/Contact.js
@@ -1,16 +1,0 @@
-import React, { Component } from "react";
- 
-class Contact extends Component {
-    render() {
-        return (
-            <div className="column middle">
-                <h2>GOT QUESTIONS?</h2>
-                <p>
-                    Hello world
-                </p>
-            </div>
-        );
-    }
-}
-
-export default Contact;
