+++ src/components/About.js
@@ -1,5 +1,5 @@
 import React, {Component} from "react";
-import './App.css'
+//import './App.css'
 
 // Router
 import {
